implement resizing text boxes

+++ src/utils/web.rs
@@ -12,11 +12,11 @@ pub async fn ask_to_persist_storage() -> Result<bool, wasm_bindgen::JsValue> {
     Ok(result.as_bool().unwrap())
 }
 
-pub fn get_screen_size() -> (u32, u32) {
+pub fn get_screen_size() -> (f64, f64) {
     let window = web_sys::window().unwrap();
     let width = window.inner_width().unwrap().as_f64().unwrap();
     let height = window.inner_height().unwrap().as_f64().unwrap();
-    (width as u32, height as u32)
+    (width, height)
 }
 
 /// Convenience function to avoid repeating expect logic.
